requirements.txt
hikari
hikari-lightbulb
aiosqlite
lightbulb-ext-filament
python-dotenv